pyproject.toml
[project]
name = "backend"
version = "0.0.0"
description = "Add your description here"
authors = [{ name = "aspizu", email = "aspizu@protonmail.com" }]
dependencies = [
    "starlette>=0.37.2",
    "msgspec>=0.18.6",
    "python-dotenv>=1.0.1",
    "wsproto>=1.2.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = ["uvicorn>=0.30.1"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/backend"]

[tool.pyright]
typeCheckingMode = "strict"
reportMissingTypeStubs = false
reportUnknownArgumentType = false
reportUnknownLambdaType = false
reportUnknownMemberType = false
reportUnknownParameterType = false
reportUnknownVariableType = false
reportMissingParameterType = false

[tool.ruff.lint]
select = ['ALL']
ignore = [
    'E402',
    'D',
    'COM',
    'C408',
    'S101',
    'D203',
    'ISC001',
    'D213',
    'S311',
    'ANN101',
]
